requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-community
pymilvus
python-multipart
pydantic
python-dotenv
pypdf2
docx2txt
pandas
numpy
tiktoken